Refactor Header to the raw Header and then the composed HeaderOnline with content. Move Meta to reusable component.
