+++ src/client/components/Header.tsx
@@ -1,24 +1,17 @@
 import * as React from 'react';
-import { Link } from 'react-router-dom';
 import styled from 'styled-components';
-import Logo from './Logo';
-import Nav from './Nav';
 import HeaderContent from './HeaderContent';
 import { headerSpacing } from '../styles/vars';
 
-interface Props {
+export interface Props extends React.HTMLProps<HTMLElement> {
   className?: string;
 }
 
-export const Header: React.SFC<Props> = ({ className }) => {
+export const Header: React.SFC<Props> = props => {
+  const { className, children, ...rest } = props;
   return (
-    <header className={className}>
-      <HeaderContent>
-        <Link to="/">
-          <Logo />
-        </Link>
-        <Nav />
-      </HeaderContent>
+    <header className={className} {...rest}>
+      <HeaderContent>{children}</HeaderContent>
     </header>
   );
 };
